통계+순서수정(#13)
