+++ EMODIARY/app/src/main/res/layout/fragment_1.xml
@@ -1,28 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
+<android.support.constraint.ConstraintLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:gravity="center"
-    android:id="@+id/fragment_1">
+    android:layout_height="match_parent">
 
-    <Button
-        android:id="@+id/camera"
+    <com.github.mikephil.charting.charts.PieChart
+        android:id="@+id/piechart"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="카메라"
-        android:background="#ffe699"/>
-
-    <Button
-        android:id="@+id/album"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="앨범에서 가져오기"/>
-    
-    <ImageView
-        android:id="@+id/user_image"
-        android:layout_width="200dp"
-        android:layout_height="200dp" />
-</LinearLayout>
+        android:layout_height="match_parent"/>
+</android.support.constraint.ConstraintLayout>
